+++ src/Sql/Sql.Test/ScenarioTests/ManagedInstanceCrudScenarioTests.ps1
@@ -62,6 +62,8 @@ function Test-CreateManagedInstance
 		Assert-AreEqual $managedInstance1.TimezoneId $timezoneId
 		Assert-AreEqual $managedInstance1.PublicDataEndpointEnabled $true
 		Assert-AreEqual $managedInstance1.ProxyOverride $proxyOverride
+		Assert-AreEqual $managedInstance1.RequestedBackupStorageRedundancy $backupStorageRedundancy
+		Assert-AreEqual $managedInstance1.CurrentBackupStorageRedundancy $backupStorageRedundancy
 		Assert-AreEqual $managedInstance1.BackupStorageRedundancy $backupStorageRedundancy
  		Assert-StartsWith ($managedInstance1.ManagedInstanceName + ".") $managedInstance1.FullyQualifiedDomainName
         Assert-NotNull $managedInstance1.DnsZone
@@ -237,6 +239,41 @@ function Test-SetManagedInstance
 	finally
 	{
 		Remove-ResourceGroupForTest $rg -AsJob
+	}
+}
+
+function Test-SetRedundancy
+{
+	# Setup
+	$rg = Create-ResourceGroupForTest
+
+	try
+	{
+		$bsr = "Geo"
+
+		# Test using parameters
+		$managedInstance1 = Create-ManagedInstanceForTest $rg
+		Assert-AreEqual $managedInstance1.CurrentBackupStorageRedundancy $bsr
+		Assert-AreEqual $managedInstance1.BackupStorageRedundancy $bsr
+		Assert-AreEqual $managedInstance1.RequestedBackupStorageRedundancy $bsr
+		
+		$bsr = "Local"
+		$managedInstance2 = Set-AzSqlInstance -ResourceGroupName $rg.ResourceGroupName -Name $managedInstance1.ManagedInstanceName -BackupStorageRedundancy $bsr -Force
+
+		Assert-AreEqual $managedInstance2.CurrentBackupStorageRedundancy $bsr
+		Assert-AreEqual $managedInstance2.BackupStorageRedundancy $bsr
+		Assert-AreEqual $managedInstance2.RequestedBackupStorageRedundancy $bsr
+
+		$bsr = "Geo"
+		$managedInstance3 = Set-AzSqlInstance -ResourceGroupName $rg.ResourceGroupName -Name $managedInstance1.ManagedInstanceName -BackupStorageRedundancy $bsr -Force
+
+		Assert-AreEqual $managedInstance3.CurrentBackupStorageRedundancy $bsr
+		Assert-AreEqual $managedInstance3.BackupStorageRedundancy $bsr
+		Assert-AreEqual $managedInstance3.RequestedBackupStorageRedundancy $bsr
+	}
+	finally
+	{
+		Remove-ResourceGroupForTest $rg
 	}
 }